+++ fixture_src/ImplicitServiceDelegateType/FooServiceFactory.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Cspray\AnnotatedContainerFixture\ImplicitServiceDelegateType;
+
+use Cspray\AnnotatedContainer\Attribute\ServiceDelegate;
+
+class FooServiceFactory {
+
+    #[ServiceDelegate]
+    public static function create() : FooService {
+        return new FooService();
+    }
+
+}