+++ fixture_src/ImplicitServiceDelegateUnionType/BarService.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Cspray\AnnotatedContainerFixture\ImplicitServiceDelegateUnionType;
+
+use Cspray\AnnotatedContainer\Attribute\Service;
+
+#[Service]
+class BarService {
+
+}