+++ test/LogicalErrorApps/ImplicitServiceDelegateNoType/FooFactory.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Cspray\AnnotatedContainer\LogicalErrorApps\ImplicitServiceDelegateNoType;
+
+use Cspray\AnnotatedContainer\Attribute\ServiceDelegate;
+
+class FooFactory {
+
+    #[ServiceDelegate]
+    public static function create() {}
+
+}