+++ test/LogicalErrorApps/ImplicitServiceDelegateScalarType/FooService.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Cspray\AnnotatedContainer\LogicalErrorApps\ImplicitServiceDelegateScalarType;
+
+use Cspray\AnnotatedContainer\Attribute\Service;
+
+#[Service]
+class FooService {
+
+}